# src/Api/Swagger/View/openapi.yaml
openapi: 3.0.0
info:
  title: 'Websps API'
  version: '0.8'
paths:
...
  /api/worksheets/master/characteristics:
    get:
      tags:
        - 'Дашборд диспетчеров'
      summary: 'Получить характеристики мастеров'
      description: 'Фильтрация работает! Слишком медленно выполняется запрос к бд'
      operationId: 3ca9c06a1b36ea086c85c8acd675a616
      parameters:
        -
          name: date_start
          in: query
          description: 'Дата начальная'
          required: true
          schema:
            type: string
            format: date
          example: '2025-05-01'
        -
          name: date_end
          in: query
          description: 'Дата конечная'
          required: true
          schema:
            type: string
            format: date
          example: '2025-05-31'
        -
          name: object_ids
          in: query
          description: 'Id объектов (массив вида 1,2,3)'
          schema:
            type: string
        -
          name: employee_name
          in: query
          description: 'Подстрока имени мастера'
          schema:
            type: string
        -
          name: code_task_name
          in: query
          description: 'Код для задачи (?)'
          schema:
            type: string
        -
          name: code_task_id
          in: query
          description: 'Id кода задачи (?)'
          schema:
            type: integer
      responses:
        '200':
          description: Успешно
          content:
            application/json:
              schema: {  }
              example: "\r\n                    {\r\n    \"masters\": [\r\n        {\r\n            \"employee_id\": 727,\r\n            \"link_id\": 216,\r\n            \"report_days\": [\r\n                {\r\n                    \"day\": \"2025-05-01\",\r\n                    \"metrics\": {\r\n                        \"exists_reports_metric\": \"0.0\",\r\n                        \"count_of_good_comments\": 0,\r\n                        \"count_of_bad_comments\": 0,\r\n                        \"rebukes_count\": 0,\r\n                        \"timely_report_generation\": 1,\r\n                        \"timely_set_task\": null,\r\n                        \"fact_without_error\": null,\r\n                        \"tabulation_correction\": 1,\r\n                        \"score_master_report\": null\r\n                    }\r\n                }\r\n            ]\r\n        }\r\n    ]\r\n}\r\n                    "
  '/api/worksheets/master/characteristics/{master_id}':
    get:
      tags:
        - 'Дашборд диспетчеров'
      summary: 'Получить характеристики мастера по id'
      operationId: 2a9a13976e05c91a62692802c205bc55
      parameters:
        -
          name: date_start
          in: query
          description: 'Дата начальная'
          required: true
          schema:
            type: string
            format: date
          example: '2025-05-01'
        -
          name: date_end
          in: query
          description: 'Дата конечная'
          required: true
          schema:
            type: string
            format: date
          example: '2025-05-31'
        -
          name: master_id
          in: path
          description: 'Id мастера'
          schema:
            type: integer
      responses:
        '200':
          description: Успешно
          content:
            application/json:
              schema: {  }
              example: "\r\n                    {\r\n    \"masters\": [\r\n        {\r\n            \"employee_id\": 727,\r\n            \"link_id\": 216,\r\n            \"report_days\": [\r\n                {\r\n                    \"day\": \"2025-05-01\",\r\n                    \"metrics\": {\r\n                        \"exists_reports_metric\": \"0.0\",\r\n                        \"count_of_good_comments\": 0,\r\n                        \"count_of_bad_comments\": 0,\r\n                        \"rebukes_count\": 0,\r\n                        \"timely_report_generation\": 1,\r\n                        \"timely_set_task\": null,\r\n                        \"fact_without_error\": null,\r\n                        \"tabulation_correction\": 1,\r\n                        \"score_master_report\": null\r\n                    }\r\n                }\r\n            ]\r\n        }\r\n    ]\r\n}\r\n                    "
  /api/worksheets/master/like-from-pto-for-day:
    post:
      tags:
        - 'Дашборд диспетчеров'
      summary: 'Поставить оценку мастеру за день работ по коду задачи и объекту'
      operationId: 7fbfb187c700a19fec153d25c5d844ff
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/JsonPtoLikeMasterForDayDto'
      responses:
        '201':
          description: 'Оценка успешно поставлена'
  /api/worksheets/master/like-from-pto-for-day/comment:
    post:
      tags:
        - 'Дашборд диспетчеров'
      summary: 'Отправить комментарий к лайку/дизлайку'
      description: 'Метод сработает только для того же самого ПТОшника, который поставил оценку'
      operationId: 9c80a8c7d8447fadb14351736f2eeeb6
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/JsonPtoLikeMasterForDayDto'
      responses:
        '201':
          description: 'Комментарий успешно установлен'
...
components:
  schemas:
...
    JsonPtoLikeMasterForDayDto:
      description: 'Json для установки лайка/дизлайка мастеру за день по объекту и коду работы'
      properties:
        date:
          required:
            - true
          type: string
          format: date
          example: '2025-06-05'
        master_code_1c:
          required:
            - true
          type: string
          example: УП00024589
        score:
          description: '1 - лайк, 0 - дизлайк'
          type: integer
          example: '1'
        comment:
          type: string
          example: 'Комментарий к оценке'
      type: object
...